requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
gunicorn==21.2.0
PyMySQL==1.1.0
cryptography==41.0.4
numpy==1.26.4
pandas==2.1.1
openpyxl==3.1.2
reportlab==4.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
pypinyin==0.50.0
Pillow==10.2.0
